send.py update and bookmark command update

+++ receive.py
@@ -44,20 +44,28 @@ for u in updates :
             bot.sendMessage(chat_id = u.message.chat.id, text = '상장되지않은 회사입니다.')
             continue
         
-        # 있으면 삭제
+        # 관심종목 여부 확인
         try:
-            sql = 'DELETE FROM bookmark WHERE name = %s AND chat_id = %s'
-            inc.cursor.execute(sql, (keyword, str(u.message.chat.id)))
-            inc.conn.commit()
-            bot.sendMessage(chat_id = u.message.chat.id, text = '관심종목에"' + keyword + '"가 삭제되었습니다.')
-            
-        # 없으면 추가
+            bookmarksql = 'SELECT name FROM bookmark WHERE name = %s AND chat_id = %s'
+            inc.cursor.execute(bookmarksql, (keyword, u.message.chat.id))
+            result = inc.cursor.fetchone()
+
+            # 없으면 추가
+            if result[0] is None:
+                sql = 'INSERT INTO bookmark (name, chat_id) VALUES (%s, %s)'
+                inc.cursor.execute(sql, (keyword, u.message.chat.id))
+                inc.conn.commit()
+                bot.sendMessage(chat_id = u.message.chat.id, text = '관심종목에"' + keyword + '"가 추가되었습니다.')
+
+            # 있으면 삭제
+            else:
+                sql = 'DELETE FROM bookmark WHERE name = %s AND chat_id = %s'
+                inc.cursor.execute(sql, (keyword, str(u.message.chat.id)))
+                inc.conn.commit()
+                bot.sendMessage(chat_id = u.message.chat.id, text = '관심종목에"' + keyword + '"가 삭제되었습니다.')
+
         except:
-            inc.conn.rollback()
-            sql = 'INSERT INTO bookmark (name, chat_id) VALUES (%s, %s)'
-            inc.cursor.execute(sql, (keyword, u.message.chat.id))
-            bot.sendMessage(chat_id = u.message.chat.id, text = '관심종목에"' + keyword + '"가 추가되었습니다.')
-            inc.conn.commit()
+                inc.conn.rollback()
 
     # 도움말
     elif u.message.text == '/help' :
